// frontend/lib/generated/fastapi/schemas/publicUsersPageResult.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * ShipyardHQ API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUserSummary } from './publicUserSummary';
import type { PublicUsersPageResultNextPage } from './publicUsersPageResultNextPage';

export interface PublicUsersPageResult {
  items: PublicUserSummary[];
  total: number;
  page: number;
  pageSize: number;
  hasMore: boolean;
  nextPage?: PublicUsersPageResultNextPage;
}